Implemented repositories for CourseEnvironmentController.cs Placed method in PostHandler into PostsRepository.cs

+++ TeamWebApplication/Repositories/CommentsRepository.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.EntityFrameworkCore;
 using TeamWebApplication.Data.Database;
 using TeamWebApplication.Models;
 using TeamWebApplication.Repositories.Interfaces;
@@ -14,27 +15,63 @@ namespace TeamWebApplication.Repositories
             _db = db;
         }
 
-        public async Task DeleteCommentByIdAsync(int commentId)
+        public async Task DeleteCommentByIdAsync(int? commentId)
         {
-            Comment comment = await _db.Comments.FindAsync(commentId);
-            _db.Comments.Remove(comment);
+            if (commentId == null)
+                throw new ArgumentNullException(nameof(commentId));
+
+            var comment = await _db.Comments.FindAsync(commentId);
+
+            if (comment != null)
+            {
+                _db.Comments.Remove(comment);
+                await SaveAsync();
+            }
         }
 
-        public async Task<Comment> GetCommentByIdAsync(int commentId)
+        public async Task DeleteCommentAsync(Comment? comment)
         {
+            if (comment == null)
+                throw new ArgumentNullException(nameof(comment));
+
+            _db.Comments.Remove(comment);
+            await SaveAsync();
+        }
+
+        public async Task<Comment?> GetCommentByIdAsync(int? commentId)
+        {
+            if (commentId == null)
+                throw new ArgumentNullException(nameof(commentId));
+
             return await _db.Comments.FindAsync(commentId);
         }
 
-        public async Task<IEnumerable<Comment>> GetCommentsByCourseIdAsync(int courseId)
+        public async Task<IEnumerable<Comment>> GetCommentsByCourseIdAsync(int? courseId)
         {
+            if (courseId == null)
+                throw new ArgumentNullException(nameof(courseId));
+
             return await _db.Comments
                 .Where(comment => comment.CourseId == courseId)
+                .OrderByDescending(comment => comment.CreationTime)
                 .ToListAsync();
         }
 
-        public async Task InsertCommentAsync(Comment comment)
+        public async Task InsertCommentAsync(Comment? comment)
         {
+            if (comment == null)
+                throw new ArgumentNullException(nameof(comment));
+
             await _db.Comments.AddAsync(comment);
+            await SaveAsync();
+        }
+
+        public async Task UpdateCommentAsync(Comment? comment)
+        {
+            if (comment == null)
+                throw new ArgumentNullException(nameof(comment));
+
+            _db.Comments.Update(comment);
             await SaveAsync();
         }
 
@@ -42,12 +79,5 @@ namespace TeamWebApplication.Repositories
         {
             await _db.SaveChangesAsync();
         }
-
-        public async Task UpdateCommentAsync(Comment comment)
-        {
-            _db.Entry(comment).State = EntityState.Modified;
-            await Task.CompletedTask;
-        }
-
     }
 }
